requirements.txt
requests>=2.28.0
pyyaml>=6.0
matplotlib>=3.5.0
